requirements.txt
aiodns==2.0.0
aiohttp==3.7.0
aiohttp-socks==0.5.5
alembic==1.4.3
apache-airflow==1.10.12
apispec==1.3.3
appdirs==1.4.4
argcomplete==1.12.1
async-timeout==3.0.1
attrs==19.3.0
Babel==2.8.0
beautifulsoup4==4.9.3
bs4==0.0.1
cached-property==1.5.2
cachetools==4.1.1
cattrs==1.0.0
cchardet==2.1.6
certifi==2020.6.20
cffi==1.14.3
chardet==3.0.4
click==7.1.2
colorama==0.4.4
colorlog==4.0.2
configparser==3.5.3
croniter==0.3.35
cssselect==1.1.0
defusedxml==0.6.0
dill==0.3.2
dnspython==2.0.0
docutils==0.16
elasticsearch==7.9.1
email-validator==1.1.1
facebook-scraper==0.2.13
-e git+https://github.com/mobolic/facebook-sdk.git@b7703c2e88dab74075175ded95cd7e244b86c0ba#egg=facebook_sdk
fake-useragent==0.1.11
Flask==1.1.2
Flask-Admin==1.5.4
Flask-AppBuilder==2.3.4
Flask-Babel==1.0.0
Flask-Caching==1.3.3
Flask-JWT-Extended==3.24.1
Flask-Login==0.4.1
Flask-OpenID==1.2.5
Flask-SQLAlchemy==2.4.4
flask-swagger==0.2.14
Flask-WTF==0.14.3
funcsigs==1.0.2
future==0.18.2
geographiclib==1.50
geopy==2.0.0
google-api-core==1.23.0
google-api-python-client==1.12.5
google-auth==1.22.1
google-auth-httplib2==0.0.4
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.52.0
googletransx==2.4.2
graphviz==0.14.2
gspread==3.6.0
gunicorn==20.0.4
html2text==2020.1.16
httplib2==0.18.1
idna==2.10
iso8601==0.1.13
itsdangerous==1.1.0
Jinja2==2.11.2
json-merge-patch==0.2
jsonschema==3.2.0
lazy-object-proxy==1.5.1
lockfile==0.12.2
lxml==4.6.1
Mako==1.1.3
Markdown==2.6.11
MarkupSafe==1.1.1
marshmallow==2.21.0
marshmallow-enum==1.5.1
marshmallow-sqlalchemy==0.24.0
multidict==5.0.0
natsort==7.0.1
numpy==1.19.2
oauth2client==4.1.3
oauthlib==3.1.0
pandas==1.1.3
parse==1.18.0
pendulum==1.4.4
prison==0.1.3
protobuf==3.13.0
psutil==5.7.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycares==3.1.1
pycparser==2.20
pyee==7.0.4
Pygments==2.7.1
PyJWT==1.7.1
pyppeteer==0.2.2
pyquery==1.4.1
pyrsistent==0.17.3
PySocks==1.7.1
python-daemon==2.2.4
python-dateutil==2.8.1
python-editor==1.0.4
python-nvd3==0.15.0
python-slugify==4.0.1
python-socks==1.1.0
python3-openid==3.2.0
pytz==2020.1
pytzdata==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-html==0.10.0
requests-oauthlib==1.3.0
rsa==4.6
schedule==0.6.0
setproctitle==1.1.10
six==1.15.0
soupsieve==2.0.1
SQLAlchemy==1.3.20
SQLAlchemy-JSONField==0.9.0
SQLAlchemy-Utils==0.36.8
tabulate==0.8.7
tenacity==4.12.0
text-unidecode==1.3
thrift==0.13.0
tqdm==4.50.2
tweepy==3.9.0
-e git+https://github.com/twintproject/twint.git@ae5e7e1189be1cf319bbd55b921aca6bfb899f8c#egg=twint
tzlocal==1.5.1
unicodecsv==0.14.1
uritemplate==3.0.1
urllib3==1.25.11
w3lib==1.22.0
websockets==8.1
Werkzeug==0.16.1
WTForms==2.3.3
yarl==1.6.2
zope.deprecation==4.4.0
